+++ primitives/src/lib.rs
@@ -50,4 +50,4 @@ pub struct NFTDetails {
 /// How NFT IDs are encoded.
 pub type NFTId = u32;
 /// How the NFT series id is encoded.
-pub type NFTSeriesId = u64;
+pub type NFTSeriesId = u32;
